requirements.txt
tensorflow==2.17.1
fastapi==0.68.0
python-multipart==0.0.5
pillow==9.2.0
uvicorn==0.15.0
mlflow==1.27.0
requests==2.28.1
